Handle error 400 from pipeline/edge service
Previous to this commit, after such error, pipeline state was stuck in loading state.
Extra Reducer was added in pipelineSlice, to revert the state loading=false after the error.

Additional logic was required in StepsSection, to abort adding a new edge in the graphic element in case of an error.